Title: 1까지 나누기, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240701/1까지 나누기/divide-by-1.cpp
@@ -7,7 +7,7 @@ int main() {
     int i=1;
     while (true){
         cnt++;
-        if (double(n)/i <= 1){
+        if (n/i <= 1){
             break;
         }
         n /= i;
